ssh-client: fix uncaught error in connection disconnect
If conn emits an error, it _may_ emit another error (DISCONNECT)
in conn.end(). The second error will be uncaught (because we're
busy throwing the first one), which crashes the process.

So now we end() the connection politely only if things go well,
and destroy() it rudely in the case of error / retraction.